--- a/src/error.rs
+++ b/src/error.rs
@@ -19,6 +19,8 @@ pub enum Error {
     InvalidDiffLeaf,
     InvalidDiffNode,
     InvalidDiffPendingUpdates,
+    InvalidRebaseNode,
+    InvalidRebaseLeaf,
     AddToDiffError,
     BuilderExpectedLeaf,
     BuilderStackEmptyMerge,