--- a/src/list.rs
+++ b/src/list.rs
@@ -4,6 +4,7 @@ use crate::interface::{ImmList, Interface, MutList};
 use crate::interface_iter::{InterfaceIter, InterfaceIterCow};
 use crate::iter::Iter;
 use crate::serde::ListVisitor;
+use crate::tree::RebaseAction;
 use crate::update_map::MaxMap;
 use crate::utils::{arb_arc, int_log, opt_packing_depth, updated_length, Length};
 use crate::{Arc, Cow, Error, Tree, UpdateMap};
@@ -245,6 +246,25 @@ impl<T: TreeHash + PartialEq + Clone + Decode + Encode, N: Unsigned, U: UpdateMa
     List<T, N, U>
 {
     pub fn rebase(&self, base: &Self) -> Result<Self, Error> {
+        let mut rebased = self.clone();
+        rebased.rebase_on(base)?;
+        Ok(rebased)
+    }
+
+    pub fn rebase_on(&mut self, base: &Self) -> Result<(), Error> {
+        match Tree::rebase_on(&self.interface.backing.tree, &base.interface.backing.tree)? {
+            RebaseAction::EqualReplace(replacement) => {
+                self.interface.backing.tree = replacement.clone();
+            }
+            RebaseAction::NotEqualReplace(replacement) => {
+                self.interface.backing.tree = replacement;
+            }
+            _ => (),
+        }
+        Ok(())
+    }
+
+    pub fn rebase_via_diff(&self, base: &Self) -> Result<Self, Error> {
         // Diff self from base.
         let diff = ListDiff::compute_diff(base, self)?;
 
@@ -253,12 +273,6 @@ impl<T: TreeHash + PartialEq + Clone + Decode + Encode, N: Unsigned, U: UpdateMa
         diff.apply_diff(&mut new)?;
 
         Ok(new)
-    }
-
-    pub fn rebase_on(&mut self, base: &Self) -> Result<(), Error> {
-        let rebased = self.rebase(base)?;
-        *self = rebased;
-        Ok(())
     }
 }