This PR implements a specialised `rebase_on` rather than using diffs to rebase. It is between 5-45x faster than the diff approach according to benchmarks.

It tries harder _not_ to recurse unnecessarily, by checking subtree equality using the hash _and_ the length. The length is required because packed leaves of different lengths can have the same hash due to 0 padding (e.g. we can't differentiate between 3x `0u64` and 4x `0xu64`).

It's also faster because it avoids intermediate allocations for the diff (a bunch of btree nodes).